pyproject.toml
[project]
name = "doc-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-deepseek>=0.1.3",
    "langchain-google-genai>=2.0.10",
    "openai>=1.77.0",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
